+++ src/main/java/de/unistuttgart/iste/gits/content_service/persistence/entity/AssessmentMetadataEmbeddable.java
@@ -1,12 +1,9 @@
-package de.unistuttgart.iste.gits.content_service.persistence.dao;
+package de.unistuttgart.iste.gits.content_service.persistence.entity;
 
 import de.unistuttgart.iste.gits.generated.dto.SkillType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.util.List;
