+++ src/main/java/de/unistuttgart/iste/gits/content_service/persistence/entity/UserProgressDataEntity.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.gits.content_service.persistence.dao;
+package de.unistuttgart.iste.gits.content_service.persistence.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
