+++ src/main/java/de/unistuttgart/iste/gits/content_service/persistence/mapper/StageMapper.java
@@ -1,6 +1,6 @@
 package de.unistuttgart.iste.gits.content_service.persistence.mapper;
 
-import de.unistuttgart.iste.gits.content_service.persistence.dao.StageEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.StageEntity;
 import de.unistuttgart.iste.gits.generated.dto.Stage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
