+++ src/main/java/de/unistuttgart/iste/gits/content_service/persistence/mapper/UserProgressDataMapper.java
@@ -1,8 +1,8 @@
 package de.unistuttgart.iste.gits.content_service.persistence.mapper;
 
 import de.unistuttgart.iste.gits.common.event.UserProgressLogEvent;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.ProgressLogItemEmbeddable;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.UserProgressDataEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.ProgressLogItemEmbeddable;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.UserProgressDataEntity;
 import de.unistuttgart.iste.gits.generated.dto.UserProgressData;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
