+++ src/main/java/de/unistuttgart/iste/gits/content_service/persistence/repository/StageRepository.java
@@ -1,7 +1,7 @@
 package de.unistuttgart.iste.gits.content_service.persistence.repository;
 
-import de.unistuttgart.iste.gits.content_service.persistence.dao.ContentEntity;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.StageEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.ContentEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.StageEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
