+++ src/test/java/de/unistuttgart/iste/gits/content_service/TestData.java
@@ -1,16 +1,11 @@
 package de.unistuttgart.iste.gits.content_service;
 
-import de.unistuttgart.iste.gits.content_service.persistence.dao.AssessmentEntity;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.AssessmentMetadataEmbeddable;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.ContentMetadataEmbeddable;
-import de.unistuttgart.iste.gits.content_service.persistence.dao.MediaContentEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.entity.*;
 import de.unistuttgart.iste.gits.generated.dto.ContentType;
 import de.unistuttgart.iste.gits.generated.dto.SkillType;
 
 import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.HashSet;
-import java.util.UUID;
+import java.util.*;
 
 public class TestData {
